requirements.txt
argparse
accelerate
requests
sentencepiece
soundfile
ollama
openai-whisper
sounddevice 
